backend/app/requirement.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# OCR Engine
easyocr>=1.7.2
torch>=2.6.0
torchvision>=0.21.0

# Speech-to-Text (Whisper AI)
openai-whisper>=20240930
pydub>=0.25.1
ffmpeg-python>=0.2.0
# audioop was dropped from stdlib in Python 3.13; this backfill is needed for pydub
audioop-lts>=0.2.0

# Image Processing
Pillow>=10.4.0
numpy>=1.26.4
opencv-python>=4.10.0

# Language Detection & Translation
langdetect>=1.0.9
googletrans>=4.0.0

# Utilities
python-multipart>=0.0.9
aiofiles>=24.1.0

# Logging and Monitoring
python-json-logger>=2.0.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Environment Variables
python-dotenv>=1.0.1

# Fact-Checking System
google-genai>=0.2.0         # New Gemini API (replaces google-generativeai)
scikit-learn>=1.3.0         # Cosine similarity calculation
langid>=1.1.6               # Language identification

# Web Scraping for Content Fetching
requests>=2.31.0            # HTTP requests
beautifulsoup4>=4.12.0      # HTML parsing
lxml>=4.9.0                 # Fast HTML parser

# Media Analysis (Stage 0)
imagehash>=1.0.8            # Perceptual hashing (pHash)
psycopg2-binary>=2.9.9      # PostgreSQL adapter
serpapi>=0.1.5              # SerpAPI for Google reverse image search